libcamera: pipeline: Add Simple pipeline
This add a generic pipeline handler for simple pipelines. It currently
handles pipelines where there's only a PHY node and a sensor node and
the kernel sets up the graph.

The current implementation can deal with sunxi sun6i-csi pipelines but
other simple pipelines can be added to the infos array.

Signed-off-by: Martijn Braam <martijn@brixit.nl>