qcam: viewfinder_gl: Add shader to render packed RAW12 formats
The shader supports all 4 packed RAW12 variants.
Simple bi-linear filtering is implemented.
The 4 LS bits of the 12-bit colour values are dropped as the RGBA
format we convert into has only 8 bits per colour.

The texture coordinates passed to the fragment shader are ajusted
to point to the nearest pixel in the image. This prevents artifacts
when the image is scaled from the frame resolution to the window size.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>