[WIP] viewfinder_gl: 1st attempt to align the pixels/textels coordinates
When the viefinder window is resized, the coordinates in the texture of
the points to draw can be close to pixels boundaries. This can lead to
mismatch between raw_even and col_even values, and the the texel the
colour values are read from. To prevent image artifacts we need to
ensure that the even/odd raw/column information is in sync with the
texel coordinates.

This implementation calculates the pixel and texel coordinates of
the point to draw in the fragment shader. So the matching vertex shader
is the identity one.