qcam: viewfinder_gl: bayer_8.frag: compute the coordinates manually
Use the identity vertex shader, and compute the coordinates in the
fragment shader without using the hardware interpolator.
The texture coordinates passed to the fragment shader are ajusted
to point to the nearest pixel in the image. This prevents artifacts
when the image is scaled from the frame resolution to the window size.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>